+++ packages/partials/concurrent/src/index.ts
@@ -14,10 +14,5 @@
  * limitations under the License.
  */
 
-export * from './Classes/AbortTimeoutController';
-export * from './Classes/FiberController';
-export * from './Classes/PromiseController';
-export * from './Classes/BackgroundRunner';
-export * from './Functions/WithTimeout';
-export * from './Functions/AutoRetry';
-export type * from './Typings';
+export * from './Classes/ThrottleController';
+export * from './Classes/DebounceController';